+++ exasol_data_science_utils_python/preprocessing/scikit_learn/sklearn_contrast_matrix_transformer.py
