+++ exasol_data_science_utils_python/preprocessing/scikit_learn/sklearn_min_max_scalar.py
